Initial commit for new Proofreading Interface
This is a largely-functional, but not complete, rewrite for a new
proofreading interface to replace both the standard and the enhanced
interfaces. The new interface is built on top of the Quill JS rich
text editor.

Quill is also used in the quizzes to provide a consistent UX between
them and the standard page interface.

Code for the old interfaces is still there, as are the relevant database
records for proofreading profiles, but only the new proofreading
interface is active and wired into the site -- for proofreading pages
and the quizzes.

The work in this commit is almost entirely from Ray Papworth, despite
having Casey Peel on the commit message.